refactor: TimeStamp 패키지 이동
