+++ backend/src/main/java/turip/favoriteplace/domain/FavoritePlace.java
@@ -11,7 +11,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.SQLRestriction;
-import turip.domain.TimeStamp;
+import turip.common.domain.TimeStamp;
 import turip.favoritefolder.domain.FavoriteFolder;
 import turip.place.domain.Place;
