+++ backend/src/main/java/turip/member/domain/Member.java
@@ -11,7 +11,7 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import org.hibernate.annotations.SQLDelete;
 import org.hibernate.annotations.SQLRestriction;
-import turip.domain.TimeStamp;
+import turip.common.domain.TimeStamp;
 
 @Getter
 @Entity
